[datatypes] Refactor vectors into own file
